requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==1.11.4
django-widget-tweaks==1.4.12
importlib-metadata==4.11.3
Markdown==3.3.6
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
zipp==3.8.0
